refactor: reorganize query test suites
