+++ crates/jazz-tools/tests/query/recursive_queries.rs
@@ -1,17 +1,15 @@
 #![cfg(feature = "test")]
 
-mod support;
-
 use std::time::Duration;
 
+use crate::support::{
+    QueryRows, TestingClient, collect_stream_deltas, has_added, wait_for_rows,
+    wait_for_subscription_update,
+};
 use jazz_tools::row_input;
 use jazz_tools::server::TestingServer;
 use jazz_tools::{
     ColumnType, JazzClient, ObjectId, QueryBuilder, Schema, SchemaBuilder, TableSchema, Value,
-};
-use support::{
-    QueryRows, TestingClient, collect_stream_deltas, has_added, wait_for_rows,
-    wait_for_subscription_update,
 };
 
 const READY_TIMEOUT: Duration = Duration::from_secs(30);
@@ -135,6 +133,58 @@ fn sorted_team_names(rows: &QueryRows) -> Vec<String> {
         .collect::<Vec<_>>();
     names.sort();
     names
+}
+
+/// Verifies that a recursive gather query seeds on a matching row and
+/// transitively reaches all ancestor rows by following the edge table.
+///
+/// Actors and flow:
+///
+/// alice writes leaf -> mid -> root in `team_edges`
+/// bob subscribes to the recursive query from leaf and sees all three teams
+#[tokio::test]
+async fn recursive_gather_query_returns_seed_and_ancestors_from_edge_table() {
+    let clients = Clients::start(team_graph_schema()).await;
+    let query = QueryBuilder::new("teams")
+        .filter_eq("name", Value::Text("leaf".to_string()))
+        .with_recursive(|r| {
+            r.from("team_edges")
+                .correlate("child_team", "_id")
+                .select(&["parent_team"])
+                .hop("teams", "parent_team")
+                .max_depth(10)
+        })
+        .build();
+
+    let mut stream = clients
+        .bob
+        .subscribe(query.clone())
+        .await
+        .expect("subscribe to recursive gather query");
+    let mut log = Vec::new();
+
+    let root_id = create_team(&clients.alice, "root", None).await;
+    let mid_id = create_team(&clients.alice, "mid", None).await;
+    let leaf_id = create_team(&clients.alice, "leaf", None).await;
+    create_team_edge(&clients.alice, leaf_id, mid_id).await;
+    create_team_edge(&clients.alice, mid_id, root_id).await;
+
+    wait_for_subscription_update(
+        &mut stream,
+        &mut log,
+        QUERY_TIMEOUT,
+        "recursive gather add delta",
+        |log| has_added(log, leaf_id),
+    )
+    .await;
+
+    let rows = wait_for_rows(&clients.bob, query, "recursive gather rows", |rows| {
+        (sorted_team_names(&rows) == vec!["leaf", "mid", "root"]).then_some(rows)
+    })
+    .await;
+    assert_eq!(sorted_team_names(&rows), vec!["leaf", "mid", "root"]);
+
+    clients.shutdown().await;
 }
 
 /// Verifies that recursive gather can use a scalar column frontier and dedupe a
